split device-specific procedure: phase 1
