+++ src/badge/device/b1248.rs
@@ -1,0 +1,1 @@
+// TODO Not Implement
